Add split management and various keybindings.
Added ability to use <leader>hjk or l to move to a split, or create one if it
doesn't exist. <leader>HJKL to move a split in that direction. <leader>sr to
rotate splits. <leader>x or sc closes splits. Arrow keys manipulate the size of
splits, with up/right making the split larger, down/left making the split smaller.